+++ packages/terra-clinical-detail-view/src/DetailList.jsx
@@ -1,6 +1,7 @@
 import React, { Children, useContext } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
+import { v4 as uuidv4 } from 'uuid';
 import styles from './DetailList.module.scss';
 import DetailListItem from './DetailListItem';
 import { HeadingLevelContext } from './HeadingLevelContext';
@@ -10,6 +11,9 @@ const cx = classNames.bind(styles);
 const propTypes = {
   /**
    * The detail view list title to be displayed above the list item(s).
+   *
+   * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue) It is critical to screen reader users that the title prop be set to a value that best describes the content of the list.
+   * Screen reader users rely on the title to provide context on the list's purpose.
    */
   title: PropTypes.string,
 
@@ -45,16 +49,17 @@ const DetailList = ({
 }) => {
   const level = useContext(HeadingLevelContext);
   const HeaderLevel = `h${level}`;
+  const titleId = title ? `list-title-id-${uuidv4()}` : null;
   let titleContent;
 
   if (title) {
-    titleContent = (<HeaderLevel className={cx('title')}>{title}</HeaderLevel>);
+    titleContent = (<HeaderLevel className={cx('title')} id={titleId}>{title}</HeaderLevel>);
   }
 
   const listContent = isLabelValuePairList
-    ? <dl className={cx('list')}>{children}</dl>
+    ? <dl className={cx('list')} aria-labelledby={titleId}>{children}</dl>
     : (
-      <ul className={cx('list')}>
+      <ul className={cx('list')} aria-labelledby={titleId}>
         {Children.map(children, child => (
           child ? (
             <li key={child.id} className={cx('list-item')}>